requirements.txt
# ── LangGraph / LangChain ─────────────────────────────────────────────────────
langgraph>=0.2.28
langchain-anthropic>=0.3.0
langchain-core>=0.2.0

# ── Anthropic Claude ──────────────────────────────────────────────────────────
anthropic>=0.30.0

# ── Semgrep ───────────────────────────────────────────────────────────────────
semgrep>=1.70.0

# ── HTTP / async ──────────────────────────────────────────────────────────────
httpx>=0.27.0
aiohttp>=3.9.0
aiofiles>=23.2.1

# ── Data / parsing ───────────────────────────────────────────────────────────
pydantic>=2.7.0
PyYAML>=6.0.1
orjson>=3.10.0

# ── Reliability ───────────────────────────────────────────────────────────────
tenacity>=8.3.0

# ── CLI / output ──────────────────────────────────────────────────────────────
rich>=13.7.0
click>=8.1.7
typer>=0.12.0

# ── Utilities ─────────────────────────────────────────────────────────────────
python-dotenv>=1.0.1
hashlib2>=1.0.0   ; python_version < "3.9"
typing-extensions>=4.11.0

# ── SARIF output ──────────────────────────────────────────────────────────────
sarif-tools>=2.1.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.2.0
pytest-asyncio>=0.23.0
